% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{cBioConnect}
\alias{cBioConnect}
\title{Create a connection to c-Bioportal API}
\usage{
cBioConnect(url = "http://www.cbioportal.org/")
}
\arguments{
\item{url}{URL of the c-Bioportal website}
}
\value{
A CGDS connexion object
}
\description{
Create a connection to c-Bioportal API
}
\seealso{
\code{\link{cgdsr::CGDS}}
}

