% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{cBioDataSet}
\alias{cBioDataSet}
\title{Retrieve a dataset from c-Bioportal (data.frame cases * genetic_profile)}
\usage{
cBioDataSet(conn, profile_id, case_id,
  genes_url = "http://acsn.curie.fr/files/acsn_v1.0.gmt")
}
\arguments{
\item{conn}{A CGDS connexion to c-Bioportal}

\item{profile_id}{List of ids of the profiles we want to retrieve}

\item{case_id}{List of ids of the cases we want to retrieve}

\item{genes_url}{URL pointing to where the list of genes of interest is}
}
\value{
A list of data.frames containing the data for each gene, for all combination experiment x sample
}
\description{
Retrieve a dataset from c-Bioportal (data.frame cases * genetic_profile)
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}
\seealso{
\code{\link{importDataSet}}, \code{\link{saveDataSet}}
}

