% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{cBioNCviz}
\alias{cBioNCviz}
\title{Create NCViz object from a c-Bioportal study}
\usage{
cBioNCviz(study_id, genes_url = "http://acsn.curie.fr/files/acsn_v1.0.gmt",
  nc_url = "http://acsn.curie.fr/files/acsn_v1.0.owl", name = "",
  method = "genes")
}
\arguments{
\item{study_id}{ID of the study to retrieve.}

\item{genes_url}{URL pointing to the list of genes of interest (in .gmt format)}

\item{nc_url}{URL of the NaviCell map}

\item{name}{Name of the dataset. If not provided, the name of the study provided by cBioPortal will be used}

\item{method}{String, either "genes" or "profiles", specifying whether the data must be fetched by genes or by profiles. The result is the same, however the "genes" version is more detailed but slower and uses more memory.}
}
\value{
An NCviz object containing the data of the study
}
\description{
Retrieve data and annotations from a c-Bioportal study and select genes that are present on the NaviCell map
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}
\seealso{
\code{\link{listStudies}}, \code{\link{cBioDataSet}}, \code{\link{cBioStudy}}
}

