% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{cBioStudy}
\alias{cBioStudy}
\title{Retrieve data from a c-Bioportal study}
\usage{
cBioStudy(study_id, genes_url = "http://acsn.curie.fr/files/acsn_v1.0.gmt")
}
\description{
Retrieve data from a c-Bioportal study
}

