% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{importStudy}
\alias{importNCviz}
\alias{importStudy}
\title{Import a study from a file}
\usage{
importStudy(fname)

importNCviz(fname, nc_url = "http://acsn.curie.fr/files/acsn_v1.0.owl",
  cell_type = "guess")
}
\arguments{
\item{fname}{Name of the file}

\item{nc_url}{URL of the NaviCell server}

\item{cell_type}{Name of the data. If cell_type == "guess", it will be deduced from the file name by default.}
}
\value{
A list (indexed by profiling methods) of dataframes (genes * samples), and the annotations in a dataframe (sample_id * annotation_type).

An NCviz object.
}
\description{
Import a study from a file with its annotations. A study consists of several experiments on the same set of genes and samples.
For a study on m patients and n genes, the file must be formated as "M experiment_name", followed by a line with the list of samples ids, followed by n lines "genes_name data".
The annotations start by "ANNOTATIONS study_name", followed by a line with the list of annotations names, followed by m lines "sample_name annotation".

Import a study from a file in an NCviz object.
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}
\seealso{
\code{\link{cBioStudy}}, \code{\link{saveData}}
}

