% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\name{NCdisplay}
\alias{NCdisplay}
\title{Display data on a NaviCell map}
\usage{
NCdisplay(obj)
}
\arguments{
\item{obj}{NCviz object}
}
\description{
Display the data on a NaviCell map using an optimized visualisation setup
}

