% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\docType{methods}
\name{NCdisplay,NCviz-method}
\alias{NCdisplay,NCviz-method}
\title{Display data on a NaviCell map}
\usage{
\S4method{NCdisplay}{NCviz}(obj)
}
\arguments{
\item{obj}{NCviz object}
}
\description{
Display the data on a NaviCell map using an optimized visualisation setup
}

