% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\docType{class}
\name{NCviz-class}
\alias{NCviz}
\alias{NCviz-class}
\title{Biological profilin data visualisation}
\usage{
NCviz(nc_url = "", cell_type = "", annotations = list(),
  cbio_data = list(), nc_data = list(), verbose = TRUE)
}
\arguments{
\item{nc_url}{URL of the NaviCell server}

\item{cell_type}{Name of the dataset to pass to the NaviCell server}

\item{annotations}{Cases annotations (clinical data)}

\item{cbio_data}{Data loaded with cBioStudy (incompatible with nc_data)}

\item{nc_data}{Data under the NaviCell format, a list indexed by experiment type containing data arranged in a dataframe with genes HUGO indentifiers as rownames and samples id as colnames (incompatible with cbio_data)}
}
\description{
S4 class for biological profiling data visualisation. Contains statistics and functions to communicate with a NaviCell API.

Constructor for the class NCviz. NCviz objects provide functions to visualize profiling data on a NaviCell map
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}
\seealso{
cBioStudy
}

