% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\docType{methods}
\name{saveData}
\alias{saveData}
\title{Save the data in one files.}
\usage{
\S4method{saveData}{NCviz}(obj, path = "./", suffix = "")
}
\arguments{
\item{obj}{NCviz object}

\item{path}{Folder where the files must be save, can be used to append a prefix to the filename}

\item{suffix}{Suffix appended to the filename}
}
\value{
Produces a .txt file, no R output.
}
\description{
Save the data in a .txt files. The file cannot be directly exported to NaviCell but can be imported with the RncMapping package. For easily commucating data.
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}

