% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\name{saveInFiles}
\alias{saveInFiles}
\title{Save the data in several files.}
\usage{
saveInFiles(obj, path = "./", suffix = "")
}
\arguments{
\item{obj}{NCviz object}

\item{path}{Folder where the files must be save, can be used to append a prefix to the filename}

\item{suffix}{Suffix appended to the filename}
}
\value{
Produces .tsv files, no R output.
}
\description{
Save the data in several files, one .tsv file per profiling method. Each file can then be imported into NaviCell.
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}

