% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\name{saveInFiles}
\alias{saveInFiles}
\title{Save the data in several files, one file per profiling method. Each file can then be imported into NaviCell.}
\usage{
saveInFiles(obj)
}
\description{
Save the data in several files, one file per profiling method. Each file can then be imported into NaviCell.
}

