% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{cBioStudy}
\alias{cBioStudy}
\title{Retrieve data and annotations from a c-Bioportal study}
\usage{
cBioStudy(study_id, genes_url = "http://acsn.curie.fr/files/acsn_v1.0.gmt")
}
\arguments{
\item{study_id}{ID of the study to retrieve.}

\item{genes_url}{URL pointing to the list of genes of interest (in .gmt format)}
}
\value{
A list containing the genes data as a list of dataframe (sample_id * profiling method, list indexed by gene names), and the annotations in a dataframe (sample_id * annotation_type)
}
\description{
Retrieve data and annotations from a c-Bioportal study and select genes that are present on the NaviCell map
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}
\seealso{
\code{\link{listStudies}}, \code{\link{cBioDataSet}}, \code{\link{cBioNCviz}}
}

