% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_cBio.R
\name{listStudies}
\alias{listStudies}
\title{List of c-Bioportal studies}
\usage{
listStudies(conn = "http://www.cbioportal.org/")
}
\arguments{
\item{conn}{A CGDS connexion object}
}
\value{
A data.frame with c-Bioportal studies ids (cancer_study_id), name (name) and description (description)
}
\description{
Get the list of c-Bioportal cancer studies
}
\seealso{
\code{\link{cBioConnect}}, \code{\link{cgdsr::CGDS}}, \code{\link{cgdsr::getCancerStudies}}
}

