% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genes_data.R
\docType{methods}
\name{saveInFiles,NCviz-method}
\alias{saveInFiles,NCviz-method}
\title{Save the data in several files.}
\usage{
\S4method{saveInFiles}{NCviz}(obj, path = "./", suffix = "")
}
\arguments{
\item{obj}{NCviz object}

\item{path}{Folder where the files must be save, can be used to append a prefix to the filename}

\item{suffix}{Suffix appended to the filename}
}
\value{
Produces .tsv files, no R output.
}
\description{
Save the data from a NCviz object in several files, one .tsv file per profiling method. Each file can then be imported into NaviCell.
Also produces a file with annotations of all samples, also importable in NaviCell.
}
\author{
Mathurin Dorel \email{mathurin.dorel@curie.fr}
}

